SOLVED - Steps in Prime
This solution can be further improved into somewhere below linear time. It's space complexity can also be improved by removing the prime_list altogether. I intend to solve it first and optimize it later, here I've solved it and feeling lazy to optimize it. It leave it up to the reader to find that 'best' solution.